+++ src/NavigationStack.js
@@ -104,6 +104,15 @@ export default class NavigationStack {
     );
   }
 
+  // This function could potentially be exposed but there seems to be no use for it.
+  //
+  // getSavedScrollPositionForLocation(location, scrollableContainerKey) {
+  //   if (!this._scrollPositionRestoration) {
+  //     throw new Error('`manageScrollPosition: true` option not passed');
+  //   }
+  //   return this._scrollPositionRestoration._getSavedScrollPositionForLocation(location, scrollableContainerKey)
+  // }
+
   init(initialLocation) {
     if (this._location) {
       throw new Error('Already initialized');