+++ src/WebNavigationStack.js
@@ -1,0 +1,8 @@
+import WebBrowserEnvironment from './environment/WebBrowserEnvironment.js'
+import NavigationStack from './NavigationStack.js'
+
+export default class WebNavigationStack extends NavigationStack {
+	constructor(options) {
+		super(WebBrowserEnvironment, options)
+	}
+}