+++ test/exports/exports.test.js
@@ -5,19 +5,20 @@ import * as exports from '../../src/index.js';
 
 describe('index', () => {
 	it('should export top level correctly', () => {
-		expect(exports.addBasePath).to.exist;
-		expect(exports.removeBasePath).to.exist;
-		expect(exports.getLocationUrl).to.exist;
-		expect(exports.parseLocationUrl).to.exist;
-		expect(exports.addNavigationBlocker).to.exist;
-		expect(exports.getLocationUrl).to.exist;
-		expect(exports.parseLocationUrl).to.exist;
-		expect(exports.parseInputLocation).to.exist;
-		expect(exports.NavigationStack).to.exist;
-		expect(exports.InMemoryEnvironment).to.exist;
-		expect(exports.WebBrowserEnvironment).to.exist;
-		expect(exports.ServerSideRenderEnvironment).to.exist;
-		expect(exports.ServerSideRedirectError).to.exist;
-		expect(exports.NavigationOutOfBoundsError).to.exist;
+		expect(exports.addBasePath).to.be.a('function');
+		expect(exports.removeBasePath).to.be.a('function');
+		expect(exports.getLocationUrl).to.be.a('function');
+		expect(exports.parseLocationUrl).to.be.a('function');
+		expect(exports.addNavigationBlocker).to.be.a('function');
+		expect(exports.getLocationUrl).to.be.a('function');
+		expect(exports.parseLocationUrl).to.be.a('function');
+		expect(exports.parseInputLocation).to.be.a('function');
+		expect(new exports.NavigationStack(exports.InMemoryEnvironment)).to.exist;
+		expect(exports.default).to.be.a('function');
+		expect(exports.InMemoryEnvironment).to.be.a('function');
+		expect(exports.WebBrowserEnvironment).to.be.a('function');
+		expect(exports.ServerSideRenderEnvironment).to.be.a('function');
+		expect(exports.ServerSideRedirectError).to.be.a('function');
+		expect(exports.NavigationOutOfBoundsError).to.be.a('function');
 	});
 });